Move ScannerControls to qr-code package
